In tests/test_alarm.py, improve how check of alarm escalation cancellation is done.
There is a race condition between a background task is cancelled and the actual cancelation is done. The test now will check it the escalation_timer_task is done, if it is it will assert that it was cancelled, if not it will await for the task and assert that it raises a CancelledError.