In tests/test_csc.py, update the test_make_log_entry unit test to publish an initial summaryState event for the ATDome CSC, used during the test.
In DDS every unit test was fully isolated but this is no longer the case with Kafka. So, if we don't publish an initial summaryState the Watcher CSC ends up getting whatever last state was published for this component, which interferes with the unit test. To ensure this does not happen the best thing to do is publish a value of the summary state the CSC would be in.