In rules/mt_m1m3_temperature.py, fix issue that was causing a memory leak.
Basically the original code passes a dictionary to the method and edits the dictionary in place. One have to be really carefull when doing this, especially as one would usually not expect that a passed argument be directly edited by a method. Furthermore, when an assignement is made, a local variable is created instead, which was the source of the memory leak.